--- src/components/Nav/MenuItem/DarkMode.css
@@ -1,8 +1,0 @@
-#darkMode-btn {
-  width: 80%;
-  height: 80%;
-  background: none;
-  outline: none;
-  border: none;
-  cursor: pointer;
-}
