+++ src/components/Nav/ServiceCenter.css
@@ -1,0 +1,25 @@
+#serviceCenter-form {
+  position: fixed;
+  min-width: 370px;
+  height: 70%;
+  border: 1px solid black;
+  bottom: 20px;
+  right: 10px;
+  z-index: 9998;
+  border-radius: 20px;
+  box-sizing: border-box;
+}
+
+#serviceCenter-headerWrap {
+  width: 100%;
+  height: 20%;
+  border-top-left-radius: 20px;
+  border-top-right-radius: 20px;
+  background-color: var(--main_color);
+}
+
+#serviceCenter-inputBox {
+  resize: none;
+}
+
+/* darkmode */
